Work on team page
Modals open and close correctly now, the table will populate with users from the database correctly. Need to work on server integration for when you want to edit a users info so that the database entry changes. Also need to work on the form page, so that it clears when you submit the form